Use FFmpeg to download from the RTSP stream directly
Summary:
Simplify the interfaces and use FFmpeg to stream the video data directly to a file

It's possible to modify the `starttime` and strip all other query parameters of the RTSP url returned by the search API. The Hikvision DVR will start at (almost, probably depending on keyframes) the indicated time and stream until quit, or until an actual break in the recordings. So it does not matter that the data on the DVR consists of multiple files, as long as the files are sequantial in time.

The search API will indicate if the `starttime` starts before an actual recording or if the `starttime + duration` passes the end of the recordings and the code will adjust these parameters.

It is possible to directly pass the `duration` to the `ffmpeg` command, but when it's past the end of the recording, it will just hang. If the `duration` is not passed to `ffmpeg`, the streaming will just stop when it reaches the end of that recording or when the duration from the `ffmpeg` status line indicates is larger or equal than the given `duration`.

Test Plan: Try different scenarios and see expected results

Reviewers: #team_nl, richard

Reviewed By: #team_nl, richard

Subscribers: richard

Differential Revision: https://phabricator.expert-shops.com/D20516