Improve logging and add storage of the accesslog
Summary:
Make the logging more consistent and make it easier to debug. Also having an accesslog is good practice.
Also add the backtrace information to the response when in debug mode

Test Plan: See logging to the console when no path is specified, otherwise see the logfiles. See backtrace information in the responses when debugging is enabled and an exception occurs

Reviewers: #team_nl, kris

Reviewed By: #team_nl, kris

Differential Revision: https://phabricator.expert-shops.com/D20850