Add a callback URI to the download request
Summary: The callback will be called when the download finishes successfully or with an error. The state can be determined from the job data POSTed to the callback.

Test Plan: Use an endpoint on my dev that echo's the request and see that it is called when a job is done and job data is passed

Reviewers: #team_nl, kris

Reviewed By: #team_nl, kris

Differential Revision: https://phabricator.expert-shops.com/D20858